refactor nav hierarchy
