+++ client/src/components/navbar/ProfileMenu.js
@@ -1,20 +1,12 @@
 import * as React from 'react';
-import { Link } from 'react-router-dom';
 
+import PropTypes from 'prop-types';
 
-import { withStyles } from '@material-ui/core/styles';
-import SwipeableDrawer from '@material-ui/core/SwipeableDrawer';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import List from '@material-ui/core/List';
-import Typography from '@material-ui/core/Typography';
 import IconButton from '@material-ui/core/IconButton';
-import Button from '@material-ui/core/Button';
 import MenuItem from '@material-ui/core/MenuItem';
 import Menu from '@material-ui/core/Menu';
 import AccountCircle from '@material-ui/icons/AccountCircle';
-import Divider from '@material-ui/core/Divider';
-import MenuIcon from '@material-ui/icons/Menu';
+
 import { withRouter } from 'react-router-dom'
 
   
@@ -29,8 +21,7 @@ class ProfileMenu extends React.Component {
     }
   
     handleLogout = () => {
-        // TODO: Issue logout action 
-      this.setState({open: false})
+      this.props.logout();
     }
     render() {
       const { open } = this.state
@@ -64,5 +55,9 @@ class ProfileMenu extends React.Component {
       )
     }
   }
+
+ProfileMenu.propTypes = {
+  logout: PropTypes.func.isRequired
+}
 
 export default withRouter(ProfileMenu);
