--- client/src/components/navbar/TopLink.js
@@ -1,16 +1,0 @@
-// @flow
-import React from 'react';
-
-type Props = {
-    variant: string,
-    className: any,
-    text: string,
-}
-
-const TopLink = (props: Props) => {
-    return (
-        <div> hi</div>
-    )
-};
-
-export default TopLink
